+++ tests/cases/annotationLayer.js
@@ -514,6 +514,47 @@ describe('geo.annotationLayer', function () {
       expect(layer.mode()).toBe(null);
       layer.options('clickToEdit', false);
     });
+    it('_handleMouseClick with modifier', function () {
+      layer.removeAllAnnotations();
+      layer.mode('polygon');
+      expect(layer.annotations().length).toBe(1);
+      expect(layer.annotations()[0].options('vertices').length).toBe(0);
+      var time = Date.now();
+      layer._handleMouseClick({
+        buttonsDown: {left: true},
+        modifiers: {shift: true},
+        time: time,
+        map: {x: 10, y: 20},
+        mapgcs: map.displayToGcs({x: 10, y: 20}, null)
+      });
+      layer._handleMouseClick({
+        buttonsDown: {left: true},
+        time: time,
+        map: {x: 30, y: 20},
+        mapgcs: map.displayToGcs({x: 30, y: 20}, null)
+      });
+      layer._handleMouseClick({
+        buttonsDown: {left: true},
+        time: time + 1000,
+        map: {x: 30, y: 20},
+        mapgcs: map.displayToGcs({x: 30, y: 20}, null)
+      });
+      layer._handleMouseClick({
+        buttonsDown: {left: true},
+        time: time + 1000,
+        map: {x: 20, y: 50},
+        mapgcs: map.displayToGcs({x: 20, y: 50}, null)
+      });
+      layer._handleMouseClick({
+        buttonsDown: {left: true},
+        time: time + 1000,
+        map: {x: 20, y: 50},
+        mapgcs: map.displayToGcs({x: 20, y: 50}, null)
+      });
+      expect(layer.annotations()[0].options('vertices').length).toBe(3);
+      expect(layer.annotations()[0].state()).toBe(geo.annotation.state.done);
+      layer.removeAllAnnotations();
+    });
     it('_handleMouseOn', function () {
       var rect = geo.annotation.rectangleAnnotation({
         layer: layer,