+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 - Polygon operations ([#1198](../../pull/1198))
+- Use meta keys to modify annotations ([#1209](../../pull/1209))
 
 ## Version 1.8.10